Add count_eq functions, but don't use them for BBitMinHasher yet.
